requirements.txt
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
PyYAML==6.0.2
